+++ tardis/plasma/standard_plasmas.py
@@ -34,8 +34,8 @@ from tardis.plasma.properties.property_collections import (
     adiabatic_cooling_properties,
     two_photon_properties,
     isotope_properties,
-    nlte_solver_properties,
     nlte_lu_solver_properties,
+    nlte_root_solver_properties,
 )
 from tardis.plasma.exceptions import PlasmaConfigError
 
@@ -215,7 +215,7 @@ def assemble_plasma(config, simulation_state, atom_data=None):
             if config.plasma.nlte_solver == "lu":
                 plasma_modules += nlte_lu_solver_properties
             elif config.plasma.nlte_solver == "root":
-                plasma_modules += nlte_solver_properties
+                plasma_modules += nlte_root_solver_properties
             else:
                 raise PlasmaConfigError(
                     "NLTE solver type unknown - {}".format(
